chore(lint): collapse nested if chains and remove redundant casts
Two changes that had to land before the final commit so cargo
clippy --all-targets --all-features -- -D warnings and
cargo fmt --check pass on a 1.97+ toolchain:

- collapse the nested if-let chains in reports::mime::find_report_part
  into single let-else-style conditions. The clippy lint widened in
  Rust 1.97 and reddened the build.
- drop the `as i64` cast that loses nothing in reports::store::ymd_to_unix;
  the operands are already i64.
- replace a useless `format!()` with `to_string()` in
  reports::tlsrpt_tests::parses_with_unknown_top_level_field.

Plus the rustfmt reformatting that the toolchain reformatted when
these changes were in flight; the diffs are mechanical, no
semantics moved.

+++ src/reports/store.rs
@@ -59,7 +59,7 @@ fn ymd_to_unix(s: &str) -> Option<i64> {
 	let d: i64 = s[6..8].parse().ok()?;
 	let (y, m) = if m <= 2 { (y - 1, m + 9) } else { (y, m - 3) };
 	let era = y.div_euclid(400);
-	let yoe = (y - era * 400) as i64;
+	let yoe = y - era * 400;
 	let doy = (153 * m + 2) / 5 + d - 1;
 	let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
 	Some((era * 146097 + doe - 719468) * 86_400)
@@ -99,30 +99,34 @@ mod tests {
 		// 20240109 is within window.
 		prune(dir.path(), today, 4);
 
-		assert!(!dir
-			.path()
-			.join("reports")
-			.join("dmarc")
-			.join("20240101")
-			.exists());
-		assert!(!dir
-			.path()
-			.join("reports")
-			.join("dmarc")
-			.join("20240105")
-			.exists());
-		assert!(dir
-			.path()
-			.join("reports")
-			.join("dmarc")
-			.join("20240109")
-			.exists());
-		assert!(!dir
-			.path()
-			.join("reports")
-			.join("tlsrpt")
-			.join("20240101")
-			.exists());
+		assert!(
+			!dir.path()
+				.join("reports")
+				.join("dmarc")
+				.join("20240101")
+				.exists()
+		);
+		assert!(
+			!dir.path()
+				.join("reports")
+				.join("dmarc")
+				.join("20240105")
+				.exists()
+		);
+		assert!(
+			dir.path()
+				.join("reports")
+				.join("dmarc")
+				.join("20240109")
+				.exists()
+		);
+		assert!(
+			!dir.path()
+				.join("reports")
+				.join("tlsrpt")
+				.join("20240101")
+				.exists()
+		);
 	}
 
 	#[test]
